requirements.txt
streamlit
requests
python-dotenv
fpdf
PyMuPDF
python-docx
matplotlib
google-generativeai